added practice 1

+++ 3rdDocument/Program.cs
@@ -3,7 +3,7 @@ Console.Write("Введите целое число: ");
 
 int userNumber = int.Parse(Console.ReadLine() ?? "");
 
-int result = (int)Math.Pow(userNumber, 2);
+double result = (int)Math.Pow(userNumber, 2);
 
 Console.WriteLine($"{userNumber} -> {result}");
